Renamed components and add missing titles from .njk pages
