+++ src/components/account-created/account-created-controller.ts
@@ -1,9 +1,9 @@
 import { Request, Response } from "express";
 
 export function registerAccountCreatedGet(req: Request, res: Response): void {
-  res.render("register-account-created/index.njk");
+  res.render("account-created/index.njk");
 }
 
 export function registerAccountCreatedPost(req: Request, res: Response): void {
-  res.render("register-account-created/index.njk");
+  res.render("account-created/index.njk");
 }
