+++ src/components/check-your-phone/check-your-phone-routes.ts
@@ -1,4 +1,4 @@
-import { basicMiddlewarePipeline } from "../../middleware/middleware-pipeline";
+import { validateSessionMiddleware } from "../../middleware/session-middleware";
 import { PATH_NAMES } from "../../app.constants";
 import express from "express";
 import { checkYourPhoneGet } from "./check-your-phone-controller";
@@ -7,7 +7,7 @@ const router = express.Router();
 
 router.get(
   PATH_NAMES.CHECK_YOUR_PHONE,
-  basicMiddlewarePipeline,
+  validateSessionMiddleware,
   checkYourPhoneGet
 );
