+++ src/components/create-password/index.njk
@@ -4,6 +4,8 @@
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
 {% from "govuk/components/details/macro.njk" import govukDetails %}
 
+{% set pageTitleName = 'pages.createPassword.title' | translate %}
+
 {% block content %}
 
 {% include "common/errors/errorSummary.njk" %}
