+++ src/components/enter-email/index.njk
@@ -4,6 +4,8 @@
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
 {% from "govuk/components/inset-text/macro.njk" import govukInsetText %}
 
+{% set pageTitleName = 'pages.enterEmail.title' | translate %}
+
 {% block content %}
 
 {% include "common/errors/errorSummary.njk" %}
