+++ src/components/enter-email/tests/enter-email-controller.test.ts
@@ -39,7 +39,7 @@ describe("enter email controller", () => {
       const fakeUserAuthService: AuthenticationServiceInterface = {
         userExists: sandbox.fake.returns(true),
         signUpUser: sandbox.fake(),
-        logInUser: sandbox.fake()
+        logInUser: sandbox.fake(),
       };
 
       req.body.email = "test.test.com";
@@ -58,7 +58,7 @@ describe("enter email controller", () => {
       const fakeUserAuthService: AuthenticationServiceInterface = {
         userExists: sandbox.fake.returns(false),
         signUpUser: sandbox.fake(),
-        logInUser: sandbox.fake()
+        logInUser: sandbox.fake(),
       };
 
       const fakeNotificationService: NotificationServiceInterface = {
@@ -85,7 +85,7 @@ describe("enter email controller", () => {
       const fakeUserAuthService: AuthenticationServiceInterface = {
         userExists: sandbox.fake.throws(error),
         signUpUser: sandbox.fake(),
-        logInUser: sandbox.fake()
+        logInUser: sandbox.fake(),
       };
 
       req.body.email = "test.test.com";
@@ -104,7 +104,7 @@ describe("enter email controller", () => {
       const fakeUserAuthService: AuthenticationServiceInterface = {
         userExists: sandbox.fake.returns(""),
         signUpUser: sandbox.fake(),
-        logInUser: sandbox.fake()
+        logInUser: sandbox.fake(),
       };
 
       req.body.email = "test.test.com";
