+++ src/components/enter-password/enter-password-routes.ts
@@ -6,19 +6,19 @@ import {
   enterPasswordPost,
 } from "./enter-password-controller";
 import { validateEnterPasswordRequest } from "./enter-password-validation";
-import { basicMiddlewarePipeline } from "../../middleware/middleware-pipeline";
+import { validateSessionMiddleware } from "../../middleware/session-middleware";
 
 const router = express.Router();
 
 router.get(
   PATH_NAMES.ENTER_PASSWORD,
-  basicMiddlewarePipeline,
+  validateSessionMiddleware,
   enterPasswordGet
 );
 
 router.post(
   PATH_NAMES.ENTER_PASSWORD,
-  basicMiddlewarePipeline,
+  validateSessionMiddleware,
   validateEnterPasswordRequest(),
   enterPasswordPost()
 );
