+++ src/components/enter-phone-number/enter-phone-number-controller.ts
@@ -1,9 +1,9 @@
 import { Request, Response } from "express";
 
 export function enterPhoneNumberGet(req: Request, res: Response): void {
-  res.render("register-enter-phone-number/index.njk");
+  res.render("enter-phone-number/index.njk");
 }
 
 export function enterPhoneNumberPost(req: Request, res: Response): void {
-  res.render("register-enter-phone-number/index.njk");
+  res.render("enter-phone-number/index.njk");
 }
