+++ src/components/enter-phone-number/enter-phone-number-routes.ts
@@ -3,21 +3,21 @@ import { PATH_NAMES } from "../../app.constants";
 import {
   enterPhoneNumberGet,
   enterPhoneNumberPost,
-} from "./register-enter-phone-number-controller";
-import { basicMiddlewarePipeline } from "../../middleware/middleware-pipeline";
-import { validateEnterPhoneNumberRequest } from "./register-enter-phone-number-validation";
+} from "./enter-phone-number-controller";
+import { validateSessionMiddleware } from "../../middleware/session-middleware";
+import { validateEnterPhoneNumberRequest } from "./enter-phone-number-validation";
 
 const router = express.Router();
 
 router.get(
   PATH_NAMES.CREATE_ACCOUNT_ENTER_PHONE_NUMBER,
-  basicMiddlewarePipeline,
+  validateSessionMiddleware,
   enterPhoneNumberGet
 );
 
 router.post(
   PATH_NAMES.CREATE_ACCOUNT_ENTER_PHONE_NUMBER,
-  basicMiddlewarePipeline,
+  validateSessionMiddleware,
   validateEnterPhoneNumberRequest(),
   enterPhoneNumberPost
 );
