+++ src/components/enter-phone-number/index.njk
@@ -3,7 +3,7 @@
 {% from "govuk/components/button/macro.njk" import govukButton %}
 {% from "govuk/components/details/macro.njk" import govukDetails %}
 {% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
-
+{% set pageTitleName = 'pages.enterPhoneNumber.title' | translate %}
 {% block content %}
 
 {% include "common/errors/errorSummary.njk" %}
