+++ src/components/footer/accessibility-statement.njk
@@ -1,4 +1,8 @@
-{% extends "common/layout/base.njk" %} {% block content %}
+{% extends "common/layout/base.njk" %}
+
+{% set pageTitleName = 'pages.enterEmail.title' | translate %}
+
+{% block content %}
 
 <h1 class="govuk-heading-l govuk-!-margin-top-0 govuk-!-margin-bottom-3">
   Accessibility
