+++ src/components/landing/landing-controller.ts
@@ -1,0 +1,6 @@
+import { Request, Response } from "express";
+import { PATH_NAMES } from "../../app.constants";
+
+export function landingGet(req: Request, res: Response): void {
+  res.redirect(PATH_NAMES.ENTER_EMAIL);
+}
