--- src/middleware/middleware-pipeline.ts
@@ -1,3 +1,0 @@
-import { validateSessionMiddleware } from "./session-middleware";
-
-export const basicMiddlewarePipeline = [validateSessionMiddleware];
