+++ src/services/types/user-exists.ts
@@ -2,4 +2,4 @@ export interface UserExists {
   email: string;
   doesUserExist: boolean;
   state: string;
-};
+}
